Big day today. Since moving over the gameobjects to game. I have now implemented 8 directional shooting with a .25 of a second cool down. Collision based on shots works for the most part. Need to remake bullet mesh. Second big task was AI. I experimented with creating my own AI today and was able to create the first pattern after hitting multiple snags The first ai behaviour pattern is that the enemy will walk towards the player consistently (this was very simple) however this ai will also recognize when its grouping up.
To do this i first overloaded my collision check to allow for a check based on a position and radius of my choosing. This radius can be altered and should allow you to modify how they group up. Currently i consider grouping up when 2 enemies touch each other. This may be changed when the map gets bigger.

After hours of working around the best way i was able to rig this is that each enemy is made with a random back up position when they are created. If an enemy touches another enemy, the first enemy will return to its back up point before chasing the player again.
While returning to the back up point the enemy can not trigger collision against other enemies. It still can collide with the player. This is done to stop a chain reaction of all enemies returning to their respective back up points.

The main issue i had with this was find where to create the random position for the enemies to return to. My working version today had a std::vector of random positions that would be used when enemies collided. I knew this was a horrible idea memory and speed wise but i did not innately notice the queue it created. Basically i had to hold the position until the enemy got there. Only then could i pop it out. That cause all enemies to form a sort of line that followed a pattern. Each time the first enemy hit its destination all the rest not chasing starting heading to spot two. all in all terrible idea. The next AI will definitely just patrol between two random spots.

To summarize: Today was spent adding in shooting and ai behaviour for the first enemy.

To do: Rewatch class on events and talk to Jimmy about the event system; as i have no idea where to even start with that. I feel like i missed something big in the understanding.