Worked on Assignment 1. Game in functional but does not include events. I do not quite understand what i am supposed to do to the vertex shader to convert values to a 0 to 100 assuming that in my 15 to 15 space that Jimmy wants 0 being -15 and 100 being 15. But what values is that referring to? What values do i even need that for?
That aside today i implemented the round buttons which were a little tricker than i expected until i figured to loop through the vector backwards.
I also added a colour shift animation to the pickups they flash from black to their full colour and back to black. The is an Imgui slider that once caught, will allow you to slide through the colour shift animation.
I experimented with it going from 0 to 2 to get all bright and with it going from 1 to 2 to go from colour to bright to colour but the shift i got felt it looked the best to me.

At this point my plan is as follows. Spend all day tomorrow attempting this event system thing and maybe looking at  that vertex conversion.

However, if this is not obtainable i will clean up THIS UPLOAD and hand it in tomorrow night. By implementing what i could in time i feel my mark will be better if this is handed in on time without events than late with events.
Also i really do not have the time to continue on this another after sunday. I am currently swamped work wise across the board.

At this point i am destroyed for the day but i will get to work as well as get some answers tomorrow.