Could not stay away lol. Think i started to crack the event system i seem to have set up a template to work off of tomorrow. My only issue i i have to call Game::OnEvent in on update which makes me wonder if i am even doing this correctly to begin with however this is what i have figured out.
I need to read better.

On my 10th looks over i noticed that it says Collision checks are still onupdate. When i was taking a shower earlier the meaning of that hit me. The collision check in on update is what creates the event. then sends the event to on event which based on the event received it will then do what needs to be done to the game objects. I will complete the setting this up tomorrow. Today i just confirmed i could use my handle collision to trigger a line in on event when something collided. I managed that well enough so hopefully tomorrow goes well. I feel hopeful for the first time about this event system which makes also sort of makes me worried that i have it completely wrong. Ill find out sooner or later.