Done for today;
Everything is in working order however the player still only collides based on the middle point.

I read the slides but i do not know how to have the collision check occur before the player moves if it happens in Game.cpp. I also do not know how to calculate the destination.
I was under the belief that the player was supposed to moved freely of the tile map so i dont really know how to begin with destination. Is my character supposed to move a tile at a time? if so that still does not solve the first issue.
I really do not know what to do from here.