half the warnings fixed (the ones in enemy)
one bug fixed related to the enemy getting stuck in the search state, this was because the enemy was never hitting the tile sometimes. I turned up the distance check to .6 and it happened less but the movement was still fluid. Turned it to .7 and have not seen the bug since, movement is also still fluid